Model-Extractor service
- Extracts the 'gets' for observers, computed properties and such in a model-extractor service;
