+++ app/services/model-extractor.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-data-model-proxy/services/model-extractor';
